+++ fenix/app/src/main/java/org/mozilla/fenix/shopping/ui/ProductAnalysis.kt
@@ -30,18 +30,13 @@ import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.Brush
-import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.unit.dp
 import org.mozilla.fenix.R
-import org.mozilla.fenix.compose.ClickableSubstringLink
-import org.mozilla.fenix.compose.SwitchWithLabel
 import org.mozilla.fenix.compose.annotation.LightDarkPreview
 import org.mozilla.fenix.compose.button.SecondaryButton
-import org.mozilla.fenix.compose.parseHtml
-import org.mozilla.fenix.settings.SupportUtils
 import org.mozilla.fenix.shopping.store.ReviewQualityCheckState
 import org.mozilla.fenix.shopping.store.ReviewQualityCheckState.HighlightType
 import org.mozilla.fenix.shopping.store.ReviewQualityCheckState.OptedIn.ProductReviewState.AnalysisPresent
@@ -101,21 +96,16 @@ fun ProductAnalysis(
             )
         }
 
-        ReviewQualityCheckExpandableCard(
-            title = stringResource(id = R.string.review_quality_check_explanation_title),
+        ReviewQualityInfoCard(
             modifier = Modifier.fillMaxWidth(),
-        ) {
-            ReviewQualityInfo(
-                modifier = Modifier.fillMaxWidth(),
-                onLearnMoreClick = onReviewGradeLearnMoreClick,
-            )
-        }
+            onLearnMoreClick = onReviewGradeLearnMoreClick,
+        )
 
-        SettingsCard(
-            modifier = Modifier.fillMaxWidth(),
+        ReviewQualityCheckSettingsCard(
             productRecommendationsEnabled = productRecommendationsEnabled,
             onProductRecommendationsEnabledStateChange = onProductRecommendationsEnabledStateChange,
             onTurnOffReviewQualityCheckClick = onOptOutClick,
+            modifier = Modifier.fillMaxWidth(),
         )
     }
 }
@@ -316,143 +306,6 @@ private fun HighlightTitle(highlightType: HighlightType) {
     }
 }
 
-@Composable
-private fun SettingsCard(
-    productRecommendationsEnabled: Boolean,
-    onProductRecommendationsEnabledStateChange: (Boolean) -> Unit,
-    onTurnOffReviewQualityCheckClick: () -> Unit,
-    modifier: Modifier = Modifier,
-) {
-    ReviewQualityCheckExpandableCard(
-        modifier = modifier,
-        title = stringResource(R.string.review_quality_check_settings_title),
-    ) {
-        Column {
-            Spacer(modifier = Modifier.height(8.dp))
-
-            SwitchWithLabel(
-                checked = productRecommendationsEnabled,
-                onCheckedChange = onProductRecommendationsEnabledStateChange,
-                label = stringResource(R.string.review_quality_check_settings_recommended_products),
-            )
-
-            Spacer(modifier = Modifier.height(16.dp))
-
-            SecondaryButton(
-                text = stringResource(R.string.review_quality_check_settings_turn_off),
-                onClick = onTurnOffReviewQualityCheckClick,
-            )
-        }
-    }
-}
-
-@Suppress("Deprecation")
-@Composable
-private fun ReviewQualityInfo(
-    modifier: Modifier = Modifier,
-    onLearnMoreClick: (String) -> Unit,
-) {
-    Column(
-        modifier = modifier,
-        verticalArrangement = Arrangement.spacedBy(24.dp),
-    ) {
-        val adjustedGradingText =
-            stringResource(id = R.string.review_quality_check_explanation_body_adjusted_grading)
-        // Any and all text formatting (bullets, inline substring bolding, etc.) will be handled as
-        // follow-up when the copy is finalized.
-        // Bug 1848219
-        Text(
-            text = stringResource(
-                id = R.string.review_quality_check_explanation_body_reliability,
-                stringResource(R.string.shopping_product_name),
-            ),
-            color = FirefoxTheme.colors.textPrimary,
-            style = FirefoxTheme.typography.body2,
-        )
-
-        val link = stringResource(
-            id = R.string.review_quality_check_info_learn_more_link,
-            stringResource(R.string.shopping_product_name),
-        )
-        val text = stringResource(R.string.review_quality_check_info_learn_more, link)
-        val context = LocalContext.current
-        val linkStartIndex = text.indexOf(link)
-        val linkEndIndex = linkStartIndex + link.length
-        ClickableSubstringLink(
-            text = text,
-            textStyle = FirefoxTheme.typography.body2,
-            clickableStartIndex = linkStartIndex,
-            clickableEndIndex = linkEndIndex,
-            onClick = {
-                onLearnMoreClick(
-                    // Placeholder Sumo page
-                    SupportUtils.getSumoURLForTopic(
-                        context,
-                        SupportUtils.SumoTopic.HELP,
-                    ),
-                )
-            },
-        )
-
-        ReviewGradingScaleInfo(
-            reviewGrades = listOf(
-                ReviewQualityCheckState.Grade.A,
-                ReviewQualityCheckState.Grade.B,
-            ),
-            info = stringResource(id = R.string.review_quality_check_info_grade_info_AB),
-            modifier = Modifier.fillMaxWidth(),
-        )
-
-        ReviewGradingScaleInfo(
-            reviewGrades = listOf(ReviewQualityCheckState.Grade.C),
-            info = stringResource(id = R.string.review_quality_check_info_grade_info_C),
-            modifier = Modifier.fillMaxWidth(),
-        )
-
-        ReviewGradingScaleInfo(
-            reviewGrades = listOf(
-                ReviewQualityCheckState.Grade.D,
-                ReviewQualityCheckState.Grade.F,
-            ),
-            info = stringResource(id = R.string.review_quality_check_info_grade_info_DF),
-            modifier = Modifier.fillMaxWidth(),
-        )
-
-        Text(
-            text = remember(adjustedGradingText) { parseHtml(adjustedGradingText) },
-            color = FirefoxTheme.colors.textPrimary,
-            style = FirefoxTheme.typography.body2,
-        )
-    }
-}
-
-@Composable
-private fun ReviewGradingScaleInfo(
-    reviewGrades: List<ReviewQualityCheckState.Grade>,
-    info: String,
-    modifier: Modifier = Modifier,
-) {
-    Row(
-        modifier = modifier.semantics(mergeDescendants = true) {},
-        verticalAlignment = Alignment.Top,
-        horizontalArrangement = Arrangement.spacedBy(8.dp),
-    ) {
-        reviewGrades.forEach { grade ->
-            ReviewGradeCompact(grade = grade)
-        }
-
-        if (reviewGrades.size == 1) {
-            Spacer(modifier = Modifier.width(24.dp))
-        }
-
-        Text(
-            text = info,
-            color = FirefoxTheme.colors.textPrimary,
-            style = FirefoxTheme.typography.body2,
-        )
-    }
-}
-
 private fun HighlightType.toHighlight() =
     when (this) {
         HighlightType.QUALITY -> Highlight.QUALITY
@@ -539,24 +392,6 @@ private fun ProductAnalysisPreview() {
                     productRecommendationsEnabled.value = it
                 },
                 onReviewGradeLearnMoreClick = {},
-            )
-        }
-    }
-}
-
-@Composable
-@LightDarkPreview
-private fun ReviewQualityInfoPreview() {
-    FirefoxTheme {
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(color = FirefoxTheme.colors.layer1)
-                .padding(all = 16.dp),
-        ) {
-            ReviewQualityInfo(
-                modifier = Modifier.fillMaxWidth(),
-                onLearnMoreClick = {},
             )
         }
     }
